Budowanie struktury plików
